backend/requirements.txt
Django==3.0.7
djangorestframework==3.11.0
python-dotenv==0.14.0